Fix after review - move mock to testscommon.
