+++ process/block/metablock_test.go
@@ -10,7 +10,6 @@ import (
 	"time"
 
 	"github.com/ElrondNetwork/elrond-go/core"
-	coreMock "github.com/ElrondNetwork/elrond-go/core/mock"
 	"github.com/ElrondNetwork/elrond-go/data"
 	"github.com/ElrondNetwork/elrond-go/data/block"
 	"github.com/ElrondNetwork/elrond-go/data/blockchain"
@@ -75,7 +74,7 @@ func createMockMetaArguments() blproc.ArgMetaProcessor {
 			BlockChain:         createTestBlockchain(),
 			BlockSizeThrottler: &mock.BlockSizeThrottlerStub{},
 			Indexer:            &mock.IndexerMock{},
-			TpsBenchmark:       &coreMock.TpsBenchmarkMock{},
+			TpsBenchmark:       &testscommon.TpsBenchmarkMock{},
 			Version:            "softwareVersion",
 		},
 		SCDataGetter:                 &mock.ScQueryStub{},
