+++ testscommon/tpsBenchmarkMock.go
@@ -1,12 +1,11 @@
-package mock
+package testscommon
 
 import (
 	"math/big"
 	"sync"
 
-	"github.com/ElrondNetwork/elrond-go/data"
-
 	"github.com/ElrondNetwork/elrond-go/core/statistics"
+	"github.com/ElrondNetwork/elrond-go/data"
 	"github.com/ElrondNetwork/elrond-go/data/block"
 )
